+++ seed/java-model/object/src/main/java/com/seed/object/model/Type.java
@@ -61,7 +61,7 @@ public final class Type {
 
     private final Name nineteen;
 
-    private final long twenty;
+    private final int twenty;
 
     private final long twentyone;
 
@@ -92,7 +92,7 @@ public final class Type {
             List<Map<String, Integer>> sixteen,
             List<Optional<UUID>> seventeen,
             Name nineteen,
-            long twenty,
+            int twenty,
             long twentyone,
             float twentytwo,
             BigInteger twentythree,
@@ -220,7 +220,7 @@ public final class Type {
     }
 
     @JsonProperty("twenty")
-    public long getTwenty() {
+    public int getTwenty() {
         return twenty;
     }
 
@@ -367,7 +367,7 @@ public final class Type {
     }
 
     public interface TwentyStage {
-        TwentyoneStage twenty(long twenty);
+        TwentyoneStage twenty(int twenty);
     }
 
     public interface TwentyoneStage {
@@ -474,7 +474,7 @@ public final class Type {
 
         private Name nineteen;
 
-        private long twenty;
+        private int twenty;
 
         private long twentyone;
 
@@ -610,7 +610,7 @@ public final class Type {
 
         @java.lang.Override
         @JsonSetter("twenty")
-        public TwentyoneStage twenty(long twenty) {
+        public TwentyoneStage twenty(int twenty) {
             this.twenty = twenty;
             return this;
         }
